Change api url, refactor, add log out
