+++ project/Mobile/package.json
@@ -13,20 +13,20 @@
     "@react-navigation/stack": "^6.4.1",
     "axios": "^1.7.7",
     "chess.js": "^1.0.0-beta.8",
-    "expo": "^51.0.37",
-    "expo-blur": "~13.0.2",
-    "expo-status-bar": "~1.12.1",
-    "react": "18.2.0",
-    "react-native": "0.74.5",
+    "expo": "^52.0.9",
+    "expo-blur": "~14.0.1",
+    "expo-network": "~7.0.0",
+    "expo-status-bar": "~2.0.0",
+    "react": "18.3.1",
+    "react-native": "0.76.2",
     "react-native-chessboard": "^0.1.2",
-    "react-native-gesture-handler": "^2.20.0",
+    "react-native-gesture-handler": "^2.20.2",
     "react-native-popover-view": "^5.1.7",
-    "react-native-reanimated": "~3.10.1",
-    "expo-network": "~6.0.1"
+    "react-native-reanimated": "^3.16.1"
   },
   "devDependencies": {
     "@babel/core": "^7.24.0",
-    "@types/react": "~18.2.79",
+    "@types/react": "~18.3.12",
     "typescript": "~5.3.3"
   },
   "private": true,
